void methods always return void.
There is a shortcut in the Nu method handlers.
Handlers for all Nu methods are inserted into the
Objective-C runtime.  But when a Nu method is called
from Nu, there is no need to go through the runtime
(and libffi) to make the call, we can just call the
method block directly.  But in these cases, methods
that are declared void will have return values,
since every Nu expression has a value.  This turned
out to be a disturbing behavior, so in the Nu-calling-Nu
"bypass" code, we now check the runtime for the
return value of the method being called. If it is
void, we force the return value of the method to be
nil.