Experimental support for escape sequences in strings.
This change set adds support for escape sequences in strings.
Escape sequences are supported in both quoted and here-strings.

These sequences are supported:
\n newline (0x0a)
\r carriage return (0x0d)
\f formfeed (0x0c)
\b backspace (0x08)
\a bell (0x07)
\e escape (0x1b)
\s space (0x20)
\nnn octal byte
\xnn hexadecimal byte
\unnnn unicode character (16 bits)
\x character x

These sequences were considered but are currently not supported:
\cx, \C-x control-x
\M-x meta-x (c | 0x80)
\M-\C-x Meta-Control-x

Because at least two existing Nu tools (nuke and nudoc) are broken
by this change, escape sequencing is currently off by default.
To enable it, use this *temporary* class method:
	(NuParser setParseEscapes:YES)

The problem is that regular expressions also use the backslash character,
so that regular expressions created from strings must use a double backslash.
Because this is unpleasant, I'm considering adding a literal form for
regular expressions in Nu, probably by wrapping the regular expression
in forward slashes.  This would replace the regex operator.

Also in this change set: the Nu parser was rewritten to be UTF-8 safe,
and exceptions thrown during parsing now cause the parser to be reset.