Character Literals for Nu.
This change set adds support for character literals in Nu.

Character literals are single characters or escape sequences
enclosed in pairs of single quotes (').  The same escape
sequences are supported for character literals and strings.
Internally, a character literal is represented by an NSNumber
with the appropriate unicode value.  This is the same value
and type that is returned by the NSString characterAtIndex:
method.

A new test was added to show a variety of character literals.

Testing revealed that the test assertion macros were not
hygenic.  This change set contains changes to fix this;
name assignments inside the test assertion macros now use
generated symbols.

Also, a few Nu tools were modified to use the new character
literal syntax.