Explicitly escaped and non-escaped forms for strings.
With this change set, strings may be specified in both
escaped and non-escaped forms.

By default, strings with no annotation ignore escape
sequences.  In the next minor release (0.2.0), this
will change and unannotated strings will expand escape
sequences.  This behavior can be controlled with the
temporary class method (NuParser setParseEscapes:).

Strings that are prefixed with a plus sign (+) explicitly
enable escaping. For example,
   +"foo\nbar"
is parsed as a seven-character string with a newline
in the middle.

Strings that are prefixed with a minus sign (-) explicitly
disable escaping.  For example,
   -"foo\nbar"
is parsed as an eight character string with a slash and
'n' in the middle.

Here strings that begin with the "<<-" notation explicitly
disable escaping. The following example shows an unescaped
string:
(set x <<-END
foo\nbarEND)

Here strings that begin with the "<<+" notation explicitly
enable escaping.  The following example shows an escaped
string:
(set x <<+END
foo\nbarEND)

There is no ambiguous form for here strings.

See the tests in test/test_strings.nu for many examples.