Closures on bindings.
With this change set, Nu closures now close on variable bindings
rather than values.  This gives Nu programmers the ability to
write functions like the accumulator-creation function in
Paul Graham's "Revenge of the Nerds" essay (for details, see
test/test_closures.nu).  It also allows for the implementation
of object-oriented programming using closures, but the value
of this is questionable, as such object-oriented systems will
be difficult to use from Objective-C.

Beginning with Nu-0.2.0, this will be the default behavior
for Nu.  Since the necessary changes were few, the original
behavior is still available with a compile-time define,
CLOSE_ON_VALUES.

The new scoping rules exposed a bug in nuke caused by name
conflicts; assignment to a variable named "task" was overriding
the definition of the "task" macro.  That has been fixed in
this change set.

See the test/test_closures.nu tests for some simple
demonstrations.