v0.2.0, literals and closures.
This release contains changes that may affect existing
Nu programs.  First, the parser now expands character
escape sequences in strings.  This is now the default
behavior for double-quoted strings. To turn off literal
expansion for a double-quoted string, prefix it with
a minus sign, like this:
   -"this string\ndoes not\sexpand escape sequences"
whereas
   "this string\ndoes\sexpand escape sequences"

For symmetry, this is also available:
   +"this string\ndoes\sexpand escape sequences"

Similar rules apply for here strings.  Here strings
defined with the <<- mark do not expand escape
sequences, and strings defined with <<+ do.  There
is no ambiguous form for here strings.

The release also contains new literal forms for
characters and regular expressions.  Character
literals are enclosed in single quotes and may
include escape sequences.  Regular expressions are
enclosed in forward slashes and may be optionally
followed by a sequence of lower-case letters specifying
options.  The supported options are:
  i  case insensitive comparison
  s  dot metacharacter also matches newlines
  x  allow whitespaces and comments in pattern
  l  lazy quantifiers become greedy and vice versa
  m  caret and dollar anchors match at newlines

Another significant change that may affect existing Nu scripts
is that closures are now made on bindings rather than values.
This brings closures in Nu in line with the commonly-accepted
definition of "closures" in Lisp and enables some interesting
tricks, such as the ability to implement objects with closures.
It also exposed a few name-related bugs. An earlier change
set contains fixes for a problem in nuke; this change set
includes fixes for a similar problem in nudoc.  If you have
a habit of reusing variable names at different levels of scoping,
be careful; that will probably get you in trouble with the new
binding behavior.  Assignments in a nested scope will be made in
an outer scope if a binding to the same name exists in that outer
scope.

To avoid confusion with gensyms, the automatically defined names
for classes and methods are now "_class" and "_method" instead
of "__class" and "__method".  See the testMetaVariables test
in test/test_classes.nu for an example using these names.