Merged MacPorts build patch from Kevin Ballard.
Merged the patch Kevin posted on the mailing list,
and made a similar change to the Rakefile.  Now
the installation and libraries directory can be
specified with the PREFIX enviroment variable.

The nu tools were updated to use /usr/bin/env to
find nush.

A small compiler warning was fixed.