Screen shots for NuRocks, reverting a change to nubile.
This change set adds a screen shot capability to NuRocks,
just press 's' when the game window is the key window to
open a save sheet and save a jpeg of the current game view.

It also reverts a change to nubile... sorry Kevin,
but setting #! to /usr/bin/env in nubile causes my
invocations from TextMate to fail.  I'm switching
back to #!/usr/local/bin/nush until this can be
fixed.