Varargs for blocks. Operators to create arrays and dictionaries.
With these changes, blocks (and functions) can be declared
to accept variable numbers of arguments.  To do this, the
last parameter in the parameter list must begin with an
asterisk character ("*").  When the body of the block is
evaluated, this name (including the "*") will be assigned
to a list consisting of the remaining function arguments.
As usual, all arguments to the block will be evaluated.

This change set also contains two applications of the
new varargs capability: "array" and "dict" operators,
written in Nu, that convert their arguments into
NSMutableArrays and NSMutableDictionaries.  "array" converts
its arguments into an array, and "dict" treats its arguments
as key-value pairs for the dictionary to create.  In the
special case where labels are given as keys, their string
values (excluding the final colon) are used as the dictionary
keys.

Tests for these changes are in test/test_varargs.nu.