Automatic key accessors for dictionaries, better testing for falsehood.
This branch contains a new handleUnknownMessage:withContext: method
for NSDictionaries that converts unknown messages into calls to
objectForKey: with the message as the key.

It also contains a small improvement to the if/unless/while/until
operators: previously the values they tested were considered false
if they were nil or had an integer value of zero.  This caused
floating point values between zero and one to be undesirably treated
as false.  Now we test that the doubleValue is zero; in the future
it may be preferable to test equality with [NSNumber numberWithInt:0],
but that will require an upgrade to the arithmetic operators, which
currently coerce all of their operands (and results) to doubles.

One more small change: nonexistent calls to [Class name] were
replaced with direct accessors of the class name field.  In the future,
this should be replaced with calls into the new objc runtime that
is provided with OS 10.5.