TextMate bundle overhaul.
This change set contains an overdue cleanup of the language grammar
in the TextMate bundle.  Syntax highlighting for strings, character
literals, and regular expressions looks much better now, and some
extraneous patterns in the grammar were removed.  Also, a firstLineMatch
pattern and an additional fileType was added to cause nush scripts and
Nukefiles to be recognized as Nu source files.

To do:
* highlight method declarations, including those specified
  with the +/- operators.
* review/fix entity names.