Added support for invocation-based undo.
In Cocoa, undo operations are implemented using the NSUndoManager class.
Via a clever "trick", messages can be sent to NSUndoManager instances
and remembered for future execution.  The "trick" is that the Objective-C
runtime can automatically convert unsupported messages into instances of
the NSInvocation class.  These NSInvocations can be saved and optionally
played out later.

Nu now supports the automatic creation of NSInvocations when unrecognized
messages are sent from Nu.  This makes it possible to use Nu to set up
undo actions and to work with any other classes that rely on NSInvocations
to capture and forward messages.