Cleaned up most of the Leopard compile warnings.
Now all of the compilation warnings on Leopard are due
to one function, now called:
   class_addInstanceVariable_withSignature()

This function adds an instance variable to a class
at runtime, after the underlying class object has
been constructed.  This is a scary thing to do,
especially if you've already created instances of
that class (if so, you should be very afraid).

But Nu is more about giving you (me) a sharp edge
than protecting us from them, so we go for it.