Leopard compile warnings rise again.
I found an error in the main Nukefile that was preventing the 10.5 SDK from explicitly being included.
When I fixed it, these warning messages reappeared for any file that included /usr/include/objc/objc-runtime.h:

In file included from /usr/include/objc/objc-runtime.h:1,
                 from objc/objc_runtime.h:12,
                 from objc/class.m:6:
/usr/include/objc/runtime.h:372: warning: ‘objc_ivar’ is deprecated (declared at /usr/include/objc/runtime.h:39)
/usr/include/objc/runtime.h:390: warning: ‘objc_method’ is deprecated (declared at /usr/include/objc/runtime.h:38)
/usr/include/objc/runtime.h:428: warning: ‘Symtab’ is deprecated
/usr/include/objc/runtime.h:454: warning: ‘objc_method_list’ is deprecated (declared at /usr/include/objc/runtime.h:53)
/usr/include/objc/runtime.h:455: warning: ‘objc_method_list’ is deprecated (declared at /usr/include/objc/runtime.h:53)
/usr/include/objc/runtime.h:460: warning: ‘objc_method’ is deprecated (declared at /usr/include/objc/runtime.h:38)

The only way to get rid of them appears to be to define __OBJC2__, which completely removes the objc structure definitions.
Not very open, and not flexible enough for Nu.