nil ([NSNull null]) is now bridged as zero.
To simplify the bridging between Nu and Objective-C, nil is now converted to a
zero-value pointer when it is passed from Nu to Objective-C.  In the past, this
was not done because the standard container classes (NSArray, NSDictionary, and
NSSet) are unable to contain zero-value pointers.  Nu uses them throughout its
implementation, especially relying on NSDictionary to store variable bindings,
and since it is important to be able to assign a variable to nil, we represent
nil in Nu with [NSNull null].

This change converts [NSNull null] to a zero-value pointer now when the bridge
is crossed.  As a result, calls to [NSDictionary setObject:forKey:] now receive
a zero-valued pointer when they should be getting [NSNull null].  This would
be a problem if not for our ability to use method-swizzling to make the container
classes more resilient.

The objc/swizzles.m file contains replacements for object insertion methods
for NSArray, NSDictionary, and NSSet that are swapped in at runtime.  These
replacements automatically convert zero-valued pointers into [NSNull null]
when they are received to be stored in the container.  Because these three
classes are implemented with underlying Core Foundation classes, the method
swizzles are performed on those classes: NSCFArray, NSCFDictionary, and NSCFSet.
To get the necessary methods on these classes, the methods are first defined
on auxiliary classes (for example, NSCFArraySwizzles), and are then copied to
the Core Foundation class using Nu's class inclusion capabilities.

With these changes, the NuZero class is no longer needed and will be removed
soon.  All occurrences of the symbols "ZERO" and "NULL" should now be replaced
with "nil".  "nil" can now safely be passed to Objective-C methods, where it will
be interpreted as a zero-valued pointer, which is exactly as it is in Objective-C.