v0.2.3, automatic invocations and a new meaning for nil.
This is a relatively small release but contains two notable
new features.  The first is that messages sent to objects
from Nu are automatically converted to NSInvocations and
forwarded when they are received by objects that can handle
them (as determined by a call to methodSignatureForSelector:).
With this change, invocation-based undo is now available
to Nu programs.

The second change eliminates the need for "NuZero", the
special class of object used to pass zero-valued pointers
to Objective-C code.  It was necessary because in Nu,
nil is represented by [NSNull null].  That is because
nil is frequently stored in Foundation collection classes,
such as NSDictionaries, and Foundation collections cannot
contain zero-valued pointers.  That, it turns out, is
a somewhat artificial restriction.  If we modify selected
mutator methods to accept zero-valued pointers and
convert them to [NSNull null], then we can modify the Nu
bridge to Objective-C to always convert [NSNull null]
(aka "nil") to zero when it crosses the bridge.  That is
what is done in this release, thanks to the flexibility
of the Objective-C runtime and a technique called "method
swizzling".  With this change set, NuZero instances are
no longer needed but the class has not yet been deleted.
It will disappear in v0.3.0.

This release also contains a new C-style ternary operator
and code changes (many submitted by Patrick Thompson) that
reduce the number of compiler warnings that are generated
by builds on Mac OS 10.5.