v0.2.4, BridgeSupport, pass-by-reference, min/max operators, and more.
This release contains preliminary support for the BridgeSupport
files that Apple included in OS 10.5.  To use it, first load
the bridgesupport code with
  (load "bridgesupport")
then import any desired frameworks with the new import macro:
  (import Cocoa)
The reader is a quick-and-easy implementation that is not tuned for
speed, so it is probably too slow for production use.  Use it to
look up symbol values to use in your application.

This release also contains changes that allow us to write
methods that return objects by reference.  Use the new (id *)
argument type in method declarations.  The arguments will then
be bridged as instances of the NuReference class.

Other changes include new min and max operators, an enhancement
to the "+" operator to allow string concatenation, a somewhat
less dirty hack to turn off an annoying and innocuous warning
message, and a new file "fscript.nu" that contains a function
that loads the F-Script browser from Nu (requiring F-Script,
of course).  Use (fs-browse myObject) to browse your favorite
objects.