API extensions.
This change set adds two new methods to the parser interface
in the Nu public header file (Nu/Nu.h).  They are valueForKey:
and setValue:forKey:.  You can use them to get and set variable
values in the top-level execution context.  valueForKey: is
more general and allows any arbitrary Nu expression to be
used as a key.  It is evaluated and the result is returned
as the value.

Also, a few new functions were added to make it possible
to programmitically construct Nu code objects in external
Objective-C code.  These functions are experimental and
may change or disappear in the future.  They are also
declared in Nu/Nu.h.

I also fixed an error in nuke.  The "exit()" function
was bridged incorrectly.