New tool "nubake" bakes Nu source code into Objective-C.
This change set adds a new tool "nubake" to the Nu tools
directory. nubake is a surprisingly simple script that
reads a Nu source file and generates an Objective-C
source file.  This source file contains a function
that, when called, returns the same parsed code
structure that is created when the original file is
parsed.  This allows Nu source code to be "baked into"
compiled Objective-C programs and libraries.

Also, for consistency, the "enu" tool has been renamed to
"nutemplate".