Improved BridgeSupport interface.
With these changes, the Nu BridgeSupport file reader is now written
in Objective-C and is fast enough for regular use.  It also includes
support for the BridgeSupport dynamic libraries that provide implementations
of functions that are inlined in Objective-C and thus are missing from
the regular framework libraries.  BridgeSupport is enabled by default
but is not automatically used.  To import the BridgeSupport files for
a framework, use the "import" macro:
  (import Foundation) - loads the Foundation framework
  (import Cocoa) - loads all of Cocoa
In both of the above cases, the loaded frameworks depend on other frameworks,
and these dependencies are listed in their BridgeSupport files.  As a result,
they can be (and are) automatically loaded by the Nu BridgeSupport reader.

Loading Cocoa loads over 10,000 symbols.  This seems like an excessive
thing to do for a production application, but for development purposes,
it can be helpful to have all of these symbols already loaded.  For
release versions of Nu-based software, developers are advised to explicitly
declare all symbols used.