Minor cleanup of protocol support.
It is now possible to create protocols at runtime with Nu.
This is done using a set of extensions to the Protocol
class, but a simpler interface is also provided in the
new "protocol" macro.  Here are some example protocols:

(protocol ChatterServing
     (- (oneway void) sendMessage:(in bycopy id) message fromClient:(in byref id) client)
     (- (BOOL) subscribeClient:(in byref id) newClient)
     (- (void) unsubscribeClient:(in byref id) client))

(protocol ChatterUsing
     (- (bycopy id) nickname)
     (- (oneway void) showMessage:(in bycopy id) message fromNickname:(in bycopy id) nickname))

As can be seen, Objective-C type modifiers are fully supported.

There is one limitation: it is currently not possible to annotate the id
type with a protocol (e.g. "id<ChatterUsing>").

For another example, see test/test_protocols.nu.