Fixed nuke error, added some helpful methods to NuBridgeSupport.
This change set undoes an earlier change that attempted to fix builds
of applications, etc. with spaces in their names.  It worked for
single targets but failed for fat ones.

It also includes a couple of helpful methods for working with
Apple's BridgeSupport files.  [NuBridgeSupport prune] removes
all BridgeSupport declarations that are unused in a program,
and [NuBridgeSupport stringValue] generates a string containing
Nu code which will regenerate the current BridgeSupport dictionary.
Use them to reduce the startup time of production applications.