Steps toward profiling and coverage analysis.
This change set adds two new fields to NuCells that
indicate the file and number of the line that was
parsed to create each cell.  Their values are used
by dtrace probes that can be used to track list
evaluations.

This is experimental and likely to change.

In the future, the file and line fields will only be
added to cells constructed by the parser, perhaps
also only when Nu is run in a special mode.