Programmatically-constructed application menu items are now displayed in bold.
This change set fixes a small problem in the construction of menus from Nu.
Application menus should now be given the title "Application", which will
be automatically replaced with a special code that Apple uses to recognize
the Application menu and make it display in boldface.

Thanks to Patrick Thomson, Bill Hale, and Colin Barrett for researching this.