More progress on x86_64.
These changes include significant progress on x86_64.
The x86_64 Nu can successfully run nuke to rebuild
itself and nutest to run the Nu regressions.  All but
one of the Nu unit tests pass; the broken one tests
the Protocol support that used mach_override and
a few other tricks to create an API to build
protocols on the fly.

since currently there is no x86_64 port of
mach_override, we'll either have to do without
or make one ourselves.  Besides breaking the
Protocol API (which may be hard to fix since
Apple's hidden much of the Protocol
implementation details in the 64bit runtime),
this also means that the infamous
	"Did you forget to call super?"
messages are temporarily back (x86_64 only).

Due to restrictions in the 64bit ObjC runtime,
it is only possible to add instance variables to
classes in class declarations that construct new
classes (by naming their superclasses) and all
instance variables must be added to a class
before any class or instance methods are added.

If you want to try the x86_64 build, you'll
have to modify the Nukefile to include it
in the architecture list.  It's not yet built
by default.