Partial workaround for missing x86_64 mach_override.
I have a partial port of mach_override to x86_64,
but it still has a few problems (such as where to
put the branch island), so in the meantime, here's
a quick-and-dirty replacement that we can use to
continue to temporarily disable NSLog when allocated
but uninitialized views are retained.  It just
writes a ret opcode to the beginning of NSLog and
saves the old value.

Runtime code modification is fun, but supporting
resources are seriously lacking.  I'm still missing
a good opcode/mnemonic converter and a guide to
memory layout in darwin programs.