nuke climbs the directory hierarchy to the nearest Nukefile.
nuke can now be run from project subdirectories.

When it is run, it looks for a Nukefile in the current directory.
If none is found, nuke climbs up the directory hierarchy until
either the root or a directory containing a Nukefile is reached.
The process working directory is set to this directory and nuke
proceeds as if it were run from this directory.