Deprecating the ternary operator.
I recently discovered a problem with the ternary operator (x ? a : b).
Because it is implemented as a message, all of its arguments are
evaluated, which is not the usual and expected behavior for such
an operator.  Specifically, it breaks recursive functions like
the one below:
(function fact (x)
  ((eq x 0) ? 1 : (* x (fact (- x 1)))))

Because all of the arguments are evaluated, the recursive evaluation
never stops, and a call to this function will crash.

So I would like to get rid of this operator.  Right now if you
use it, you'll get a deprecation warning.