Fixed a problem reading 10.5 ivar type encodings.
It seems that beginning with Mac OS 10.5, when an instance variable
that is a pointer to an object is created by the compiler, its type encoding
includes the type of the pointed-to object.  For example, the encoding
for "NSArray *" is now @"NSArray" where previously it was @.  This change
broke some ivar management code in Nu, and that is fixed in this change
set.

I also discovered that a recent change to nuke uses features that aren't
available in mininush, and that has also been fixed.