comparison operators now use compare:
The <, >, <=, and >= operators now all rely on the
compare: method to perform the necessary tests.
They also now accept an arbitrary number of
arguments (at least one!).  Each operator returns
t (true) if its arguments are ordered appropriately
and nil (false) if not.  For example,
  (< 1 2 3) returns t, and
  (< 1 2 2) returns nil

This allows us to write functions like this one
which tests that its argument is a list of objects
sorted in increasing order:

(function increasing (mylist)
     (eval (cons '< mylist)))