Fixed memory leaks in the parser.
Thanks to Grayson Hansard, we found and fixed some sloppy memory usage in the Nu parser.
In most Nu applications (until now), the problem didn't arise, because typically a parser
would be created once at the beginning of an application and used through its lifetime.

But in Grayson's example, Nu parsers were being created, used, and discarded repeatedly
as part of embedding Nu inside an application, and we discovered that a retain cycle was
preventing these parsers from being released.  To break the retain cycle, the "close"
method should be called at the end of a parser's lifetime.

For example:

    NSAutoreleasePool *pool = [[NSAutoreleasePool alloc] init];
    // get an autoreleased parser
    id nu = [Nu parser];
    // do something with the parser
    [nu eval:[nu parse:@"YOUR CODE GOES HERE"]];
    // now close the parser
    [nu close];
    // the parser will be released with the pool
    [pool release];