Baking-in the Nu library.
With these changes, the Nu source files in nu/*.nu can be
pre-parsed and written to Objective-C source files that
can in turn be compiled and linked into Nu.  This will
allow Nu to be built as a single-file static or shared
library, which will make it easier to deploy Nu on platforms
that don't support frameworks (or at least not as well
as Mac OS X does).  Also, if you build Nu as a static
library (libNu.a, not yet directly supported in this
change set), then it can be directly linked into an
application's binary.