iPhone support. To build libNu.a, "cd iphone; nuke"
This commit also adds "baked in" versions of the
Nu source files in the nu directory.  These need
to be updated whenever the corresponding nu files
change.  Do that with "nuke bake" in the top-level
directory.  As a future enhancement, it would be
great to remove these files from the repo and
build them on demand. Also, we are currently baking
a lot of extra stuff into the iPhone libNu.a.