Linux-related build fixes.
Now both Darwin and Linux versions of Nu can be built
and installed using the Nu two-stage build process.

Stage 1: Use rake to build a statically-linked Nu shell.

	% rake

Stage 2: Use the resulting 'mininush' to run nuke.

	% mininush tools/nuke

Installation Notes: On Darwin, Nu is installed as a
framework in /Library/Frameworks. On Linux, it is built
as a dynamic library and installed in /usr/local/lib
(or the lib dir of whatever is your desired installation
prefix). Because nush is linked against libNu.so,
running "nuke install" will crash on Linux when the
libNu.so is overwritten. So be sure to install with
"mininush tools/nuke install".

Linux builds require NuFound, a customized version of
libFoundation. Ports to other Foundation libraries are
welcome.