New upTo:do: and downTo:do: methods for NSNumbers.
Based on a suggestion from Jordan Breeding, NSNumbers now
support upTo:do: and downTo:do: methods that allow iteration
over specific ranges of integers.

For example,
% (5 upTo:10 do:(do (i) (puts "-- #{i}")))
-- 5
-- 6
-- 7
-- 8
-- 9
-- 10
5

Note that the return value is the receiving object.

I don't particularly like the camel case in upTo:
and downTo:, but that seems necessary for consistency.