Try to force class +initialize methods to be called before sending messages to class object.
This change appears to fix a problem reported by Jordan Breeding. Here's the symptom:

tim% nush
Nu Shell.
% (NSPredicate predicateWithFormat: "not (%K == %@)" argumentArray: (array "kind" "PDF document"))
2008-03-26 09:47:47.380 nush[17256:807] *** -[NSRecursiveLock unlock]: lock (<NSRecursiveLock: 0x43b740> '(null)') unlocked when not locked
2008-03-26 09:47:47.382 nush[17256:807] *** Break on _NSLockError() to debug.
<NSCompoundPredicate:444cc0>

I suspect that this has something to do with the +initialize method of NSPredicate not
being called. On subsequent calls in the same process, there is no message. Also,
when I run (NSPredicate initialize) before the call, there is no message.

This one-line change appears to trigger a call to +initialize. With the
change, Jordan's example runs with no error message.