Replaced swizzled methods with methods that explicitly allow nil to be added to dictionaries, arrays, and sets.
Since contexts are represented with NSDictionaries, it is useful to be able
to insert nil values into them. These new methods automatically convert nil
to [NSNull null] objects, which are required to represent nil in Cocoa/NeXTSTEP
containers. They replace swizzled methods that made this capability available
to all dictionary (and array and set) insertions, but which did so at the
cost of an extra message send on every insertion. This requires that a different
method be used within the Nu implementation, but overall seems to be more clean.