Command line options for nuke, better option support in nush.
These changes add support for two command-line switches
to nuke:
    -v             run nuke verbosely.
    -f <Nukefile>  run with a specified Nukefile.

This required some improvements to nush, the Nu shell.
Now nush accepts switches before a file argument, but
any switches after the file argument are treated as
arguments to that file. This is the same way switches
are handled by ruby and probably many other interpreters.
An array of arguments may be obtained with a call to
  ((NuApplication sharedApplication) arguments)