select: and filter: methods now treat 0 (zero) as false values.
The cond, if, unless, etc. operators already were treating
zero as false which greatly simplifies the use of C functions
and Objective-C methods that return 0 to indicate false.
This change set extends that to also apply to the select:
and find: methods of NuEnumerable and NuCell and also adds
unit tests for both groups of enumeration methods.

While preparing the tests, I noticed that I had not added
a mapSelector: to NuCell when I recently added it to
NuEnumerable, so to correct that inconsistency a
(NuCell mapSelector:) is now included and tested.