Added a return operator.
This new operator allows early return from blocks, which
includes both functions and class and instance methods.
The return operator accepts an optional argument which
is used as the return value. If no argument is given,
nil is returned.

See the unit tests in test/test_return.nu for examples.