Better printing of strings in lists.
Strings are now displayed in a quoted and escaped form when
they are displayed as a part of a list. So now we have:

% (list "a" "b" "c")
("a" "b" "c")

where previously:

% (list "a" "b" "c")
(a b c)

I also fixed the order of arguments in some test assertion
macro invocations. The desired result should always be first.