Variable scoping change.
With this change, assignments made within methods cannot
affect symbols outside of the local class scope.

===========================
Example:

(set x 0)
(set z 0)

(class Tester is NSObject
  (set y 0)
  (+ doSomething is
    (set x 1)
    (set y 1)
    ...
    )
  (+ getY is y)
  (+ getZ is z))

(Tester doSomething)
===========================

Previously, the assignment to x in (Tester +doSomething)
would modify the top-level x since that binding was in
the lexical closure of doSomething. Now that same assignment
won't affect the top-level x because assignments are limited
to the local class scope. Variables outside the scope are
still visible, (Tester +getZ) returns the value of z at
the top-level.

(Tester +doSomething) also assigns a value to y. Because
there is an assignment to y in the class context, that
assignment is the one modified by the set operator. Thus
class variables can be created using closures. This is
consistent with previous versions of Nu. Only the behavior
regarding x is modified by these changes.

These changes also include some minor cleanup to the
(NuCell -stringValue) method.