Memory management error fixed, (NSData datawithShellCommand:), nil in containers.
The results of calls to certain functions were being over-retained.
Normally, the results of Nu functions and methods are returned
"autoreleased", or with no additional retains, as this is the
convention for Objective-C methods. In Objective-C, the only exceptions
to this are for alloc, allocWithZone, copy, copyWithZone, mutableCopy,
mutableCopyWithZone, and new. These methods return objects with an
extra retain, and this change set fixes incorrect handling of this
in Nu. When Nu implementations of these methods are called by Objective-C
code, they must return properly retained values. Previously Nu was
doing the right thing but for the wrong methods (init methods).
That caused objects created when Objective-C callers called Nu init
methods to be over-retained and leaked.

A new file of unit tests is included to test memory management.
Because these tests require an Objective-C component, a helper
class was added Nu in objc/testhelper.m.

As a convenience, a new (NSData dataWithShellCommand:) is added and
is similar to (NSString stringWithShellCommand:).

Also, because certain Nu applications were having problems inserting
nil into Cocoa containers (NSDictionary, NSArray, NSSet, and their
mutable variants), these changes reinstate the method-swizzling trick
used to make those containers accept nil and internally convert it
to (NSNull null).