Fixed instance variable leak, reviewed implementation of dealloc methods in Nu.
I realized today that despite all my talk of automatic memory management,
I had forgotten to explicitly release instance variables when their
owners were deallocated. This change set fixes that by swizzling a new
dealloc method onto NSObject that explicitly releases all instance
variables declared within Nu. In a future change, this swizzling will
only be done for the classes for which Nu instance variables have been
declared.

These changes were verified with a new test in test_memory.nu and a
small change in the behavior of the let operator. let now evaluates
its subexpressions in a dedicated autorelease pool that is released
upon completion of the let evaluation. This makes it easier to confirm
object deletion.

I also investigated a crash that occurred in the evaluation of dealloc
methods implemented with Nu. The crash occurs when the dealloc method
completes evaluation and its context is deallocated. When the context
is deallocated, the value assigned to "self" is released. But if
(super dealloc) has been called (as it should have been), then this
value is undefined and the release operation can cause a crash.
The solution is to set self to nil right before calling (super dealloc).
This is demonstrated in the new testIvarReleaseOnDealloc unit test.
With that, it should now be possible to safely write dealloc methods
with Nu. Just remember to always (set self nil) and call (super dealloc)
at the end of the method and never send retain messages to instance
variables from Nu code; that is handled automatically.