Fixed a problem in (NSObject set:).
via Paul Phillips, NuHTTP was failing with the following error:
$ ./nuhttpd
2008-05-28 15:26:20.946 nush[585:613] *** -[NuHTTPRequestHandler setAction:]: unrecognized selector sent to instance 0x3cc940
2008-05-28 15:26:20.948 nush[585:613] Terminating due to uncaught exception (below):
2008-05-28 15:26:20.948 nush[585:613] NSInvalidArgumentException: *** -[NuHTTPRequestHandler setAction:]: unrecognized selector sent to instance 0x3cc940

I believe this problem arose when I moved the set: method to Objective-C;
for any property named "action" it was calling the setAction: method
directly instead of using key value coding. This is because actions in
the target-action paradigm are represented with selectors and selectors
aren't objects.

But in cases where "action" is used as a general instance variable,
this is a problem, because there may not be a setAction: method that
is available from Objective-C. So now we only call setAction: if
the class responds to that selector.