Fixed a bug in [NuClass addClassMethod:].
The definition of #addClassMethod: was identical to
added instance methods to the receiver. This changeset
fixes the bug by passing a reference to the isa/class_pointer
of the class to add_method_to_class(). I also migrated
tests for #addClassMethod: and #addInstanceMethod: from
test_nu.rb into test_classes.nu.

These changes work on Mac OS X 10.5.2, but are untested
on other platforms.