Tighter memory management, but leaving some unresolved issues with nib file outlets.
These changes reinstate automatic releasing of "automatic"
Nu instance variables; these are the variables created
automatically when they are accessed for classes whose
definitions include the (ivars) operator.

Other instance variables are not automatically released
when their owners are deallocated, because I learned
that when they are set as outlets in nib files, their
values aren't retained.  This may require a separate
declaration type ("outlet" or "weak" instead of "id")
and until this is resolved, we potentially leak objects
assigned to declared instance variables.

An update to Nu to support GC would eliminate the need
for this, but Nu also needs to work on platforms without GC.

Also, on the subject of other platforms, the iPhone
build is now broken because (until recently at least),
the iPhone SDK does not support NSMapTables. That's
too bad, because the NSMapTable interface is really nice,
but portability is better, so I plan to replace them with
the symbol tables implemented in the st.[hc] files.